changed int to long

+++ tutorial/src/main/java/Fibonacci/Fibonacci.java
@@ -5,8 +5,8 @@ public class Fibonacci {
     public Fibonacci() {
     }
 
-    public int fib(int n){
-        int result=0;
+    public long fib(int n){
+        long result=0;
 
         if(n<0) {
             System.out.println("You entered a negative number");
@@ -16,8 +16,8 @@ public class Fibonacci {
 
         if(n==0 || n==1) return n;
             else{
-            int twoBefore=0;
-            int oneBefore=1;
+            long twoBefore=0;
+            long oneBefore=1;
 
             for(int i=2;i<=n; i++) {
                 result = twoBefore + oneBefore;
